Move compiler class file into components
