+++ src/Sulu/Bundle/RouteBundle/SuluRouteBundle.php
@@ -12,9 +12,9 @@
 namespace Sulu\Bundle\RouteBundle;
 
 use Sulu\Bundle\PersistenceBundle\PersistenceBundleTrait;
-use Sulu\Bundle\RouteBundle\DependencyInjection\RouteDefaultOptionsCompilerPass;
 use Sulu\Bundle\RouteBundle\DependencyInjection\RouteGeneratorCompilerPass;
 use Sulu\Bundle\RouteBundle\Model\RouteInterface;
+use Sulu\Component\Route\RouteDefaultOptionsCompilerPass;
 use Sulu\Component\Symfony\CompilerPass\TaggedServiceCollectorCompilerPass;
 use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -32,7 +32,9 @@ class SuluRouteBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new RouteGeneratorCompilerPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, -1024);
-        $container->addCompilerPass(new RouteDefaultOptionsCompilerPass());
+        $container->addCompilerPass(
+            new RouteDefaultOptionsCompilerPass('sulu_route.routing.provider', 5)
+        );
         $container->addCompilerPass(
             new TaggedServiceCollectorCompilerPass('sulu_route.routing.defaults_provider', 'sulu_route.defaults_provider')
         );
